💄 Style: 테이블 현황,목록 css 수정

+++ client/src/components/AdminComponents/TableStatus/TableStatus.js
@@ -1,6 +1,6 @@
 import { useSelector } from 'react-redux';
 import styled from 'styled-components';
-
+import { RiDeleteBinLine } from 'react-icons/ri';
 const TableStatus = ({ data }) => {
    console.log(data);
    const priceList = data.orderList.map(menus => {
@@ -9,10 +9,12 @@ const TableStatus = ({ data }) => {
    const url = useSelector(state => state.adminReducer.apiUrl);
    const totalPrice = priceList.reduce((prev, current) => prev + current);
    const hadleClickDeleteOrder = () => {
-      fetch(`${url}/order/${sessionStorage.getItem('userId')}/${data.tableNumber}`, {
-         method: 'DELETE',
-         headers: { 'Content-Type': 'application/json' }
-      });
+      if (confirm('정말 삭제하시겠습니까?')) {
+         fetch(`${url}/order/${sessionStorage.getItem('userId')}/${data.tableNumber}`, {
+            method: 'DELETE',
+            headers: { 'Content-Type': 'application/json' }
+         });
+      }
    };
 
    return (
@@ -21,7 +23,9 @@ const TableStatus = ({ data }) => {
             <div>No. {data.tableNumber}</div>
             <div>
                <div className="orderDelete">
-                  <button onClick={hadleClickDeleteOrder}>삭제</button>
+                  <button onClick={hadleClickDeleteOrder}>
+                     <RiDeleteBinLine size="15"></RiDeleteBinLine>
+                  </button>
                </div>
                <div>
                   주문 <b>{data.orderList.length}</b>개
@@ -56,7 +60,7 @@ const TableStatusBox = styled.div`
    box-shadow: 0 4px 2px 0px lightgray;
    font-size: 1.3rem;
    padding: 20px;
-   background-color: #fdffde;
+   background-color: #fffff3;
    border-radius: 3px 3px 15px 3px;
    margin-bottom: 25px;
    .orderDelete {
@@ -65,6 +69,9 @@ const TableStatusBox = styled.div`
       width: 100%;
       justify-content: end;
       > button {
+         display: flex;
+         align-items: center;
+         justify-content: center;
          font-size: 12px;
       }
    }
@@ -73,11 +80,11 @@ const TableStatusBox = styled.div`
       align-items: center;
       width: 100%;
       height: 80px;
-      border-bottom: 1px solid gray;
+      border-bottom: 1px solid #c2a9a9;
       font-weight: 900;
       > :nth-child(1) {
          font-size: 30px;
-         color: #ff6c01;
+         color: #ff5252;
          font-weight: bold;
          width: 50%;
          margin-left: 10px;
@@ -112,7 +119,7 @@ const TableStatusBox = styled.div`
       }
    }
    .totalPrice {
-      border-top: 1px solid gray;
+      border-top: 1px solid #c2a9a9;
       width: 100%;
       height: 10%;
       > div {
